Get successes to be green & force one test to fail
In the adder-test.html file I force one of the two tests to fail with
the fail() method just to have the two different outputs on the console.
in the tinytest.js file I put a list of new features to implement on
the app and then did the first one: Get successes to be green.